update fiber & scheduler

+++ include/fishjoy/util.hpp
@@ -19,8 +19,8 @@ namespace fishjoy
   pid_t GetThreadId();
   uint32_t GetFiberId();
 
-  void Backtrace(std::vector<std::string>& bt, int size, int skip = 1);
-  std::string BacktraceToString(int size, int skip = 2, const std::string& prefix = "");
+  void Backtrace(std::vector<std::string>& bt, unsigned long size, int skip = 1);
+  std::string BacktraceToString(unsigned long size, int skip = 2, const std::string& prefix = "");
 
 }  // namespace fishjoy
